requirements.txt
Adafruit_DHT
requests
influxdb-client